fix(cli): drain offline lifecycle spills through secure storage
